# out2/src/main/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: alberto.est.jimenez@gmail.com
    name: Alberto Jimenez
  description: Documentation for the client-accessible endpoints of the StackV orchestrator.
  title: StackV Orchestrator Developer API
  version: 0.8.2
servers:
- url: https://stackv.dev3.virnao.com/StackV-web/restapi/
...
paths:
...
  /address/pool:
    get:
      description: Retrieves all accessible address pools.
      operationId: retrievePools
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AddressPoolDTO"
                type: array
          description: List of basic address pool DTOs.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve Pools
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
    post:
      description: Creates a new address pool.
      operationId: createPool
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressPoolRequest"
      responses:
        "201":
          description: Pool created.
        "500":
          $ref: "#/components/responses/Error"
      summary: Add New Pool
      tags:
      - address
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: address
  /address/pool/{poolName}:
    delete:
      description: Deletes a specific address pool.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Pool successfully deleted.
        "500":
          $ref: "#/components/responses/Error"
      summary: Delete Pool
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
    get:
      description: Retrieves a specific address pool.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressPoolDTO"
          description: The address pool DTO.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve Pool
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
  /address/allocate/{poolName}:
    get:
      description: Retrieves all allocations from a specified address pool.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AddressAllocation"
                type: array
          description: List of basic address pool DTOs.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve Address Pool Allocations
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
    post:
      description: Allocates a new address from an address pool.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocateAddressRequest"
        description: Allocation request parameters.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: string
                type: array
          description: List of addresses allocated.
        "500":
          $ref: "#/components/responses/Error"
      summary: Allocate Address
      tags:
      - address
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: address
  /address/allocate/{poolName}/{scope}:
    delete:
      description: Remove all allocations within a pool matching a specific scope.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      - description: Scope of the allocations requested.
        example: default
        explode: false
        in: path
        name: scope
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Addresses successfully deleted.
        "500":
          $ref: "#/components/responses/Error"
      summary: Delete Allocations by Scope
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
    get:
      description: Retrieves all allocations from a specified address pool with a
        specific scope.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      - description: Scope of the allocations requested.
        example: default
        explode: false
        in: path
        name: scope
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/AddressAllocation"
                type: array
          description: List of basic address pool DTOs.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve Address Pool Allocations by Scope
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
  /address/allocate/{poolName}/{address}/{scope}:
    delete:
      description: Remove a specific allocation.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      - description: Specific address (or 'all' for every address in scope).
        example: fc00:1780::/64
        explode: false
        in: path
        name: address
        required: true
        schema:
          type: string
        style: simple
      - description: Scope of the allocation.
        example: default
        explode: false
        in: path
        name: scope
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Pool successfully deleted.
        "500":
          $ref: "#/components/responses/Error"
      summary: Delete Allocation
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
  /address/access/{pool}:
    delete:
      description: Deletes all ACLs associated with a specified address pool.
      parameters:
      - description: ID of the address pool.
        example: 31
        explode: false
        in: path
        name: pool
        required: true
        schema:
          type: number
        style: simple
      responses:
        "204":
          description: Deletion occurred without exception.
        "500":
          $ref: "#/components/responses/Error"
      summary: Clear Address Pool Access List
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
    get:
      description: Retrieves all ACLs from a specified address pool.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: pool
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/paths/~1address~1access~1%7BpoolName%7D~1%7Busername%7D/get/responses/200/content/application~1json/schema"
                type: array
          description: List of basic address pool DTOs.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve Address Pool Access List
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
  /address/access/{poolName}/{username}:
    delete:
      description: Deletes the ACL associated with a specified address pool and user.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      - description: Username to assign permissions.
        example: tester
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/paths/~1address~1access~1%7BpoolName%7D~1%7Busername%7D/get/responses/200/content/application~1json/schema"
          description: The ACL entry DTO.
        "204":
          description: Deletion not necessary. Entry not found.
        "500":
          $ref: "#/components/responses/Error"
      summary: Delete Address Pool Access List
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
    get:
      description: Retrieve the ACL entry for a specified address pool and user.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      - description: Username to assign permissions.
        example: tester
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address_Pool_ACL"
          description: The ACL entry DTO.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve Address Pool Access List
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
  /address/access/{poolName}/{username}/{level}:
    put:
      description: Updates the ACL entry for a specified address pool and user.
      parameters:
      - description: Name of the address pool.
        example: Rucio Test Pool
        explode: false
        in: path
        name: poolName
        required: true
        schema:
          type: string
        style: simple
      - description: Username to assign permissions.
        example: tester
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      - description: New permission level to set.
        example: READ
        explode: false
        in: path
        name: level
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/paths/~1address~1access~1%7BpoolName%7D~1%7Busername%7D/get/responses/200/content/application~1json/schema"
          description: The updated ACL entry DTO.
        "500":
          $ref: "#/components/responses/Error"
      summary: Update Address Pool Access List
      tags:
      - address
      x-accepts:
      - application/json
      x-tags:
      - tag: address
...
components:
  responses:
    Error:
      description: Internal server fault or standardized error response.
...
  schemas:
    AddressAllocation:
      description: This schema represents a specific address allocation.
      example:
        address: 192.168.66.15
        scope: default
        name: Dev5 IP 6
        id: 552
        expiry: 1698174979000
        affiliate: urn:ogf:network:nrp-nautilus.io:2020:sn3700_s0:PortChannel501:vlanport+3987
      properties:
        id:
          description: Unique ID of the address allocation.
          example: 552
          format: int64
          type: integer
        name:
          description: Name of the address allocation.
          example: Dev5 IP 6
          type: string
        address:
          description: The address itself.
          example: 192.168.66.15
          type: string
        expiry:
          description: The address expiration timestamp.
          example: 1698174979000
          type: number
        scope:
          description: The address scope.
          example: default
          type: string
        affiliate:
          description: Name of the address allocation.
          example: urn:ogf:network:nrp-nautilus.io:2020:sn3700_s0:PortChannel501:vlanport+3987
          type: string
      required:
      - address
      - id
      - name
      title: Address Allocation
      type: object
    AllocateAddressRequest:
      example:
        allocate:
          netmask: netmask
          scope: scope
          batch: batch
          exclusive: exclusive
        name: Rucio Test IP
        type: ID
      properties:
        name:
          description: The name of the allocation.
          example: Rucio Test IP
          type: string
        type:
          enum:
          - ID
          - MAC
          - IPv4
          - IPv6
          type: string
        allocate:
          $ref: "#/components/schemas/AllocateParams"
      type: object
    AllocateParams:
      example:
        netmask: netmask
        scope: scope
        batch: batch
        exclusive: exclusive
      properties:
        batch:
          type: string
        exclusive:
          type: string
        netmask:
          type: string
        scope:
          type: string
      type: object
    AddressPoolDTO:
      description: This schema represents the limited DTO of an address pool.
      example:
        name: ESNet East Pool 4
        pool: "fc00:3600::0/64,[...],fc00:3601::0/64"
        id: 19
        type: IPv4
      properties:
        id:
          description: Unique ID of the address pool.
          example: 19
          type: number
        name:
          description: Human readable name of the address pool.
          example: ESNet East Pool 4
          type: string
        type:
          description: The type of address a pool manages.
          enum:
          - IPv4
          - IPv6
          - ID
          - MAC
          type: string
        pool:
          description: CSV string of pool data.
          example: "fc00:3600::0/64,[...],fc00:3601::0/64"
          type: string
      required:
      - id
      - name
      - pool
      - type
      title: Address Pool
      type: object
    CreateAddressPoolRequest:
      example:
        prefix_list:
        - 10.0.0.1/24
        - 10.0.0.1/24
        prefix: prefix
        name: Rucio Dev Pool
        range: 192.168.1.1-192.168.1.255
        type: IPv4
      properties:
        name:
          description: The name of the address pool.
          example: Rucio Dev Pool
          type: string
        type:
          description: The type of addresses to allocate.
          enum:
          - ID
          - MAC
          - IPv4
          - IPv6
          example: IPv4
          type: string
        prefix:
          description: The type of addresses to allocate.
          type: string
        range:
          description: Create a pool from an address range.
          example: 192.168.1.1-192.168.1.255
          type: string
        prefix_list:
          description: Create a pool from multiple prefixes.
          items:
            example: 10.0.0.1/24
            type: string
          type: array
      type: object
...
    Address_Pool_ACL:
      description: This schema represents an access control configuration for an address
        pool.
      example:
        level: READ
        pool:
          name: ESNet East Pool 4
          pool: "fc00:3600::0/64,[...],fc00:3601::0/64"
          id: 19
          type: IPv4
        id: 54
        username: tester
      properties:
        id:
          description: Unique ID of the ACL entry.
          example: 54
          type: number
        pool:
          $ref: "#/components/schemas/AddressPoolDTO"
        username:
          description: Username being assigned permissions.
          example: tester
          type: string
        level:
          description: Permission level for this user.
          example: READ
          type: string
      required:
      - id
      - name
      - pool
      - type
      title: Address Pool ACL
      type: object
...
  securitySchemes:
    OAuth2:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/auth
          scopes:
            openid: OIDC Mode.
          tokenUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/token
      type: oauth2